LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-stereopanner-interface/no-dezippering-expected.txt

CONSOLE MESSAGE: line 26: SyntaxError: The string did not match the expected pattern.

Harness Error (FAIL), message = SyntaxError: The string did not match the expected pattern.

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [test mono input] Test StereoPanner with mono input has no dezippering 
